Feat: web, console 중복 로직 제거
