+++ src/main/java/racingcar/RacingCarApplication.java
@@ -9,5 +9,4 @@ public class RacingCarApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(RacingCarApplication.class, args);
 	}
-
 }
