+++ src/main/java/racingcar/RacingCarConsoleApplication.java
@@ -1,57 +1,34 @@
 package racingcar;
 
-import racingcar.domain.entity.RacingCar;
+import racingcar.domain.entity.RacingCars;
+import racingcar.domain.service.RacingCarGameService;
+import racingcar.infra.inmemory.InMemoryPlayResultRepository;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Random;
 import java.util.Scanner;
-import java.util.stream.Collectors;
 
 public class RacingCarConsoleApplication {
     public static void main(String[] args) {
         // 자동차 입력
         Scanner scanner = new Scanner(System.in);
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-        List<RacingCar> racingCars = Arrays.stream(scanner.nextLine().split(","))
-                .map(it -> new RacingCar(it.trim()))
-                .collect(Collectors.toList());
+        String names = scanner.nextLine();
+        RacingCars racingCars = RacingCars.from(names);
 
         // 시도 횟수 입력
         System.out.println("시도할 횟수는 몇 회인가요?");
         int count = scanner.nextInt();
 
         // 경주 시작
-        for (int i = 0; i < count; i++) {
-            playRound(racingCars);
-        }
+        RacingCarGameService racingCarGameService = new RacingCarGameService(new InMemoryPlayResultRepository());
+        racingCarGameService.playGame(racingCars, count);
 
         // 우승자 조회
-        int maxPosition = 0;
-        List<String> winners = new ArrayList<>();
-        for (RacingCar racingCar : racingCars) {
-            if (racingCar.getPosition() > maxPosition) {
-                maxPosition = racingCar.getPosition();
-                winners.clear();
-            }
-            if (racingCar.getPosition() >= maxPosition) {
-                winners.add(racingCar.getName());
-            }
-        }
-
+        List<String> winners = racingCars.getWinners();
         System.out.println();
         System.out.println("최종 우승자: " + String.join(", ", winners));
-        racingCars.forEach(it ->
+        racingCars.getRacingCars().forEach(it ->
                 System.out.println("name: " + it.getName() + ", " + "position: " + it.getPosition())
         );
-    }
-
-    private static void playRound(List<RacingCar> racingCars) {
-        Random random = new Random();
-        for (RacingCar racingCar : racingCars) {
-            int randomNumber = random.nextInt(10);
-            racingCar.move(randomNumber);
-        }
     }
 }
